requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9

# Vector Database & Embeddings
chromadb==0.4.18
pinecone-client==2.2.4
sentence-transformers==2.2.2
huggingface-hub==0.19.4
numpy<2.0.0

# Multimodal Processing
open-clip-torch==2.20.0
pillow==10.1.0
PyMuPDF==1.23.8
trafilatura==1.6.4
beautifulsoup4==4.12.2
faster-whisper==0.10.0
ffmpeg-python==0.2.0

# LLM Providers
openai==1.3.7
anthropic==0.7.8
google-cloud-aiplatform==1.38.1

# Caching & Utilities
redis==5.0.1
tenacity==8.2.3
structlog==23.2.0
ulid-py==1.1.0

# Graph & NLP
spacy==3.7.2
networkx==3.2.1

# Development & Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
httpx==0.25.2
faker==20.1.0

# Code Quality
ruff==0.1.6
black==23.11.0
isort==5.12.0
mypy==1.7.1

# Job Processing
rq==1.15.1
rq-dashboard==0.6.1

# File Upload & Processing
python-multipart==0.0.6

# Auth & Security
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# Monitoring
prometheus-client==0.19.0

# Additional Utilities
python-ulid==1.1.0
requests==2.31.0
